+++ tests/test_tabular_train.py
@@ -28,7 +28,7 @@ def test_accuracy(learn):
     assert learn.validate()[1] > 0.7
 
 def test_same_categories(learn):
-    this_tests('skip')
+    this_tests('na')
     x_train,y_train = learn.data.train_ds[0]
     x_valid,y_valid = learn.data.valid_ds[0]
     x_test,y_test = learn.data.test_ds[0]
@@ -39,7 +39,7 @@ def test_same_categories(learn):
         assert np.all(x_train.classes[key] == x_test.classes[key])
 
 def test_same_fill_nan(learn):
-    this_tests('skip')
+    this_tests('na')
     df = pd.read_csv(path/'adult.csv')
     nan_idx = np.where(df['education-num'].isnull())
     val = None
@@ -56,7 +56,7 @@ def test_normalize(learn):
     df = pd.read_csv(path/'adult.csv')
     train_df = df.iloc[0:800].append(df.iloc[1000:])
     c = 'age'
-    this_tests('skip')
+    this_tests('na')
     mean, std = train_df[c].mean(), train_df[c].std()
     for i in np.random.randint(0,799, (20,)):
         x,y = learn.data.train_ds[i]
@@ -69,7 +69,7 @@ def test_normalize(learn):
         assert np.abs(x.conts[0] - (df.loc[i, c] - mean) / (1e-7 + std)) < 1e-6
 
 def test_empty_cont():
-    this_tests('skip')
+    this_tests('na')
     df = pd.read_csv(path/'adult.csv')
     procs = [FillMissing, Categorify, Normalize]
     dep_var = 'salary'