+++ tests/test_utils_links.py
@@ -13,52 +13,52 @@ def build_nb_cells(mod_names):
 
 @pytest.mark.skip(reason="need to update")
 def test_torchvision():
-    this_tests('skip')
+    this_tests('na')
     docstr   = 'Note that `tvm` is the namespace we use for `torchvision.models`.'
     expected = 'Note that [`tvm`](https://pytorch.org/docs/stable/torchvision/models.html#torchvision.models) is the namespace we use for `torchvision.models`.'
     assert_link(docstr, expected, msg='Should match imported aliases')
 
 def test_fastai_prefix():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "functions for your application (`fastai.vision`)"
     expected = "functions for your application ([`fastai.vision`](/vision.html#vision))"
     assert_link(docstr, expected, msg='Should match keywords prefixed with fastai. See `index.ipynb`')
 
 def test_link_typedef():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "- `LayerFunc` = `Callable`\[`nn.Module`],`None`]"
     expected = "- `LayerFunc` = `Callable`\[[`nn.Module`](https://pytorch.org/docs/stable/nn.html#torch.nn.Module)],`None`]"
     assert_link(docstr, expected, modules=[torch], msg='Type definitions to torch formatted incorrectly. See fastai_typing.ipynb')
 
 def test_link_typedef_double_bt():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "- `ParamList` = `Collection`\[`nn`.`Parameter`]"
     expected = "- `ParamList` = `Collection`\[[`nn`](https://pytorch.org/docs/stable/nn.html#torch-nn).`Parameter`]"
     assert_link(docstr, expected)
 
 def test_link_inner_class_functions():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "To train your model in mixed precision you just have to call `Learner.to_fp16`, which converts the model and modifies the existing `Learner` to add `MixedPrecision`."
     expected = "To train your model in mixed precision you just have to call [`Learner.to_fp16`](/train.html#to_fp16), which converts the model and modifies the existing [`Learner`](/basic_train.html#Learner) to add [`MixedPrecision`](/callbacks.fp16.html#MixedPrecision)."
     imports = 'from fastai.callbacks.fp16 import *'
     assert_link(docstr, expected, nb_cells=[gen_notebooks.get_code_cell(imports)])
 
 def test_class_anchor():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "`DataBunch.create`, `DeviceDataLoader.proc_batch`"
     expected = "[`DataBunch.create`](/basic_data.html#DataBunch.create), [`DeviceDataLoader.proc_batch`](/basic_data.html#DeviceDataLoader.proc_batch)"
     imports = 'from fastai.basic_train import *'
     assert_link(docstr, expected, nb_cells=[gen_notebooks.get_code_cell(imports)])
 
 def test_link_class_methods():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "`ImageDataBunch.from_csv`"
     expected = "[`ImageDataBunch.from_csv`](/vision.data.html#ImageDataBunch.from_csv)"
     imports = 'from fastai.vision.data import *'
     assert_link(docstr, expected, nb_cells=[gen_notebooks.get_code_cell(imports)])
 
 def test_respects_import_order():
-    this_tests('skip')
+    this_tests('na')
     docstr   = "`learner`"
     expected = "[`learner`](/vision.learner.html#vision.learner)"
     assert_link(docstr, expected, build_nb_cells(['fastai.text', 'fastai.vision']))
@@ -67,7 +67,7 @@ def test_respects_import_order():
     assert_link(docstr, expected_text, build_nb_cells(['fastai.vision', 'fastai.text']))
 
 def test_nb_module_name_has_highest_priority():
-    this_tests('skip')
+    this_tests('na')
     # get_imported_modules.nb_module_name should have highest priority. This is the associated notebook module.
     # Ex: vision.transforms.ipynb is associated with fastai.vision.transforms
     docstr   = "`transform`"
@@ -81,7 +81,7 @@ def test_nb_module_name_has_highest_priority():
 
 @pytest.mark.skip(reason="need to update")
 def test_application_links_top_level_modules():
-    this_tests('skip')
+    this_tests('na')
     # Snippet taken from applications.ipynb
     docstr = """## Module structure
 In each case (except for `collab`), the module is organized this way:
@@ -98,7 +98,7 @@ In each case (except for [`collab`](/collab.html#collab)), the module is organiz
     assert_link(docstr, expected, msg='data, models should link to highest module. transform and learner links to first match')
 
 def test_link_vision_learner_priority():
-    this_tests('skip')
+    this_tests('na')
     # Edge case for vision.learner.ipynb
     imports = """from fastai.gen_doc.nbdoc import *
     from fastai.vision import *